+++ src/Repeated_Random.java
@@ -2,7 +2,7 @@
  * Created by Andrew on 4/20/2016.
  */
 public class Repeated_Random {
-    public int trial(int[] input_array, int max_Iter, String key) {
+    public long trial(long[] input_array, int max_Iter, String key) {
         representation r;
         if (key.equals("one")){
             r = new Rep_1(input_array);
@@ -14,8 +14,8 @@ public class Repeated_Random {
         int[] best_array = r.return_array();
         int[] new_array;
         int array_length = input_array.length;
-        int val_best = r.residue(best_array);
-        int val_new;
+        long val_best = r.residue(best_array);
+        long val_new;
         System.out.println("Testing Repeated Random...");
         for (int i = 0; i < max_Iter; i++) {
             System.out.println("Iteration: " + i);
